reflection/requirements.txt
backports.shutil-get-terminal-size==1.0.0
blessed==1.14.1
-e git://github.com/neuman/python-carteblanche@9ef6b77dd6de77f42e8df951d2a047fac42a758c#egg=carteblanche-origin_feature_dynamic_denial
decorator==4.0.10
dj-database-url==0.4.0
Django==1.10.1
django-bootstrap-form==3.2.1
django-bootstrap-pagination==1.6.1
django-etc==0.9.1
django-jsonfield==1.0.1
django-sitegate==0.10.3
django-termsandconditions==1.1.8
-e git://github.com/revsys/django-tos@03c099bae096631706790d49dfa67741cf0fa0d3#egg=django_tos-dev
enum34==1.1.6
gunicorn==19.4.5
ipython==5.0.0
ipython-genutils==0.1.0
keyring==8.7
keyrings.alt==1.1.1
lxml==3.6.4
measurement==1.8.0
myfitnesspal==1.8.6
oauthlib==1.1.2
pathlib2==2.1.0
pexpect==4.2.0
pickleshare==0.7.3
prompt-toolkit==1.0.5
psycopg2==2.6.1
ptyprocess==0.5.1
Pygments==2.1.3
PyJWT==1.4.2
python-dateutil==2.5.3
python-openid==2.2.5
python-social-auth==0.2.19
pytz==2016.6.1
requests==2.11.0
requests-oauthlib==0.6.2
simplegeneric==0.8.1
six==1.10.0
sympy==0.7.6.1
traitlets==4.2.2
wcwidth==0.1.7
wheel==0.24.0
whitenoise==2.0.6
